refact: coreg tests
