+++ mypy.ini
@@ -1,2 +1,5 @@
 [mypy]
 plugins = numpy.typing.mypy_plugin
+
+[mypy-tests.*]
+disable_error_code = no-untyped-def
